+++ src/socket.rs
@@ -159,11 +159,11 @@ impl Socket {
             // Do the connect call, this is non-blocking due to the socket setup
             let mut connect_result = unsafe {
                 nrfxlib_sys::nrf_connect(self.fd, address.as_ptr(), address.size() as u32)
-            };
+            } as isize;
 
-            const NRF_EINPROGRESS: i32 = nrfxlib_sys::NRF_EINPROGRESS as i32;
-            const NRF_EALREADY: i32 = nrfxlib_sys::NRF_EALREADY as i32;
-            const NRF_EISCONN: i32 = nrfxlib_sys::NRF_EISCONN as i32;
+            const NRF_EINPROGRESS: isize = nrfxlib_sys::NRF_EINPROGRESS as isize;
+            const NRF_EALREADY: isize = nrfxlib_sys::NRF_EALREADY as isize;
+            const NRF_EISCONN: isize = nrfxlib_sys::NRF_EISCONN as isize;
 
             if connect_result == -1 {
                 connect_result = get_last_error();
@@ -231,11 +231,11 @@ impl Socket {
 
             // Do the bind call, this is non-blocking due to the socket setup
             let mut bind_result =
-                unsafe { nrfxlib_sys::nrf_bind(self.fd, address.as_ptr(), address.size() as u32) };
+                unsafe { nrfxlib_sys::nrf_bind(self.fd, address.as_ptr(), address.size() as u32) } as isize;
 
-            const NRF_EINPROGRESS: i32 = nrfxlib_sys::NRF_EINPROGRESS as i32;
-            const NRF_EALREADY: i32 = nrfxlib_sys::NRF_EALREADY as i32;
-            const NRF_EISCONN: i32 = nrfxlib_sys::NRF_EISCONN as i32;
+            const NRF_EINPROGRESS: isize = nrfxlib_sys::NRF_EINPROGRESS as isize;
+            const NRF_EALREADY: isize = nrfxlib_sys::NRF_EALREADY as isize;
+            const NRF_EISCONN: isize = nrfxlib_sys::NRF_EISCONN as isize;
 
             if bind_result == -1 {
                 bind_result = get_last_error();
@@ -272,7 +272,7 @@ impl Socket {
             }
 
             let mut send_result = unsafe {
-                nrfxlib_sys::nrf_send(self.fd, buffer.as_ptr() as *const _, buffer.len() as u32, 0)
+                nrfxlib_sys::nrf_send(self.fd, buffer.as_ptr() as *const _, buffer.len(), 0)
             };
 
             if send_result == -1 {
@@ -282,7 +282,7 @@ impl Socket {
             #[cfg(feature = "defmt")]
             defmt::trace!("Send result {}", send_result);
 
-            const NRF_EWOULDBLOCK: i32 = -(nrfxlib_sys::NRF_EWOULDBLOCK as i32);
+            const NRF_EWOULDBLOCK: isize = -(nrfxlib_sys::NRF_EWOULDBLOCK as isize);
 
             match send_result {
                 bytes_sent @ 0.. => Poll::Ready(Ok(bytes_sent as usize)),
@@ -309,7 +309,7 @@ impl Socket {
             }
 
             let mut receive_result = unsafe {
-                nrfxlib_sys::nrf_recv(self.fd, buffer.as_ptr() as *mut _, buffer.len() as u32, 0)
+                nrfxlib_sys::nrf_recv(self.fd, buffer.as_ptr() as *mut _, buffer.len(), 0)
             };
 
             if receive_result == -1 {
@@ -319,7 +319,7 @@ impl Socket {
             #[cfg(feature = "defmt")]
             defmt::trace!("Receive result {}", receive_result);
 
-            const NRF_EWOULDBLOCK: i32 = -(nrfxlib_sys::NRF_EWOULDBLOCK as i32);
+            const NRF_EWOULDBLOCK: isize = -(nrfxlib_sys::NRF_EWOULDBLOCK as isize);
 
             match receive_result {
                 bytes_received @ 0.. => Poll::Ready(Ok(bytes_received as usize)),
@@ -355,7 +355,7 @@ impl Socket {
                 nrfxlib_sys::nrf_recvfrom(
                     self.fd,
                     buffer.as_ptr() as *mut _,
-                    buffer.len() as u32,
+                    buffer.len(),
                     0,
                     socket_addr_ptr,
                     &mut socket_addr_len as *mut u32,
@@ -369,7 +369,7 @@ impl Socket {
             #[cfg(feature = "defmt")]
             defmt::trace!("Receive result {}", receive_result);
 
-            const NRF_EWOULDBLOCK: i32 = -(nrfxlib_sys::NRF_EWOULDBLOCK as i32);
+            const NRF_EWOULDBLOCK: isize = -(nrfxlib_sys::NRF_EWOULDBLOCK as isize);
 
             match receive_result {
                 bytes_received @ 0.. => Poll::Ready(Ok((bytes_received as usize, {
@@ -405,7 +405,7 @@ impl Socket {
                 nrfxlib_sys::nrf_sendto(
                     self.fd,
                     buffer.as_ptr() as *mut _,
-                    buffer.len() as u32,
+                    buffer.len(),
                     0,
                     addr.as_ptr(),
                     addr.size() as u32,
@@ -419,7 +419,7 @@ impl Socket {
             #[cfg(feature = "defmt")]
             defmt::trace!("Sending result {}", send_result);
 
-            const NRF_EWOULDBLOCK: i32 = -(nrfxlib_sys::NRF_EWOULDBLOCK as i32);
+            const NRF_EWOULDBLOCK: isize = -(nrfxlib_sys::NRF_EWOULDBLOCK as isize);
 
             match send_result {
                 bytes_received @ 0.. => Poll::Ready(Ok(bytes_received as usize)),
